make rubocop happy

+++ spec/support/files/run_mockserver_container.sh
@@ -8,7 +8,12 @@ if [ "$status" == "7" ]; then
     apt-get install -qq docker.io -y 1>&- 2>&-
 fi
 
-docker run -d --rm -p 1080:1080 --name mockserver mockserver/mockserver 1>&- 2>&-
+id=`docker ps -q -f name=mockserver -f status=running`
+
+if [ -z "$id" ]
+then
+  docker run -d --rm -p 1080:1080 --name mockserver mockserver/mockserver 1>&- 2>&-
+fi
 
 id=`docker ps -q -f name=mockserver -f status=running`
